python3.7 compat diff --git a/murmur/include/crypt.h b/murmur/include/murmurcrypt.h similarity index 100% rename from murmur/include/crypt.h rename to murmur/include/murmurcrypt.h diff --git a/murmur/murmur.cpp b/murmur/murmur.cpp index 2a6e968..e48bc20 100755 --- a/murmur/murmur.cpp +++ b/murmur/murmur.cpp @@ -26,7 +26,7 @@  #include <iostream>  #include <sys/stat.h>  #include "unzip.h" -#include "crypt.h" +#include "murmurcrypt.h"
 #if PY_MAJOR_VERSION >= 3
 #define IS_PY3K
diff --git a/murmur/unzip.c b/murmur/unzip.c
index 9ad4766..c05b3dc 100755
--- a/murmur/unzip.c
+++ b/murmur/unzip.c
@@ -153,7 +153,7 @@ typedef struct

 #ifndef NOUNCRYPT
-#include "crypt.h"
+#include "murmurcrypt.h"
 #endif

 /* ===========================================================================
