+++ murmur/include/murmurcrypt.h
