+++ murmur/murmur.cpp
@@ -26,7 +26,7 @@
 #include <iostream>
 #include <sys/stat.h>
 #include "unzip.h"
-#include "crypt.h"
+#include "murmurcrypt.h"
 
 #if PY_MAJOR_VERSION >= 3
 #define IS_PY3K
